Corrected some problems and added guides for the streaming. Still needing to fix StreamingController

+++ Backend/Snapshot03.01/src/main/java/bg/fmi/uni/boomvox/dto/LogEventRequest.java
@@ -3,7 +3,7 @@ package bg.fmi.uni.boomvox.dto;
 import bg.fmi.uni.boomvox.enums.StreamingEventType;
 
 public record LogEventRequest(
-    int                sessionId,
+    Long               sessionId,
     StreamingEventType type,
     long               positionSec,
     Long               seekFromSec,
